banging my head at buttons

+++ src/firmware/rhythm.h
@@ -22,7 +22,7 @@
 // void rhythm_init();
 
 uint8_t rhythm_addval(int val);
-bool rhythm_oxygen(float dur_pct);
+bool rhythm_oxygen(int dur_pct);
 int rhythm_get_baseline();
 uint8_t rhythm_get_period();
 uint8_t rhythm_get_phase();
